refactor(ff-decode): deduplicate RAII guards into shared guards_inner module
