+++ crates/ff-decode/src/image/decoder_inner.rs
@@ -30,82 +30,7 @@ use ff_sys::{
 };
 
 use crate::error::DecodeError;
-
-// ── RAII guards ────────────────────────────────────────────────────────────────
-
-/// RAII guard for `AVFormatContext` to ensure proper cleanup.
-struct AvFormatContextGuard(*mut AVFormatContext);
-
-impl AvFormatContextGuard {
-    unsafe fn new(path: &Path) -> Result<Self, DecodeError> {
-        // SAFETY: Caller ensures FFmpeg is initialized and path is valid
-        let format_ctx = unsafe {
-            ff_sys::avformat::open_input(path).map_err(|e| DecodeError::Ffmpeg {
-                code: e,
-                message: format!("Failed to open file: {}", ff_sys::av_error_string(e)),
-            })?
-        };
-        Ok(Self(format_ctx))
-    }
-
-    const fn as_ptr(&self) -> *mut AVFormatContext {
-        self.0
-    }
-
-    fn into_raw(self) -> *mut AVFormatContext {
-        let ptr = self.0;
-        std::mem::forget(self);
-        ptr
-    }
-}
-
-impl Drop for AvFormatContextGuard {
-    fn drop(&mut self) {
-        if !self.0.is_null() {
-            // SAFETY: self.0 is valid and owned by this guard
-            unsafe {
-                ff_sys::avformat::close_input(&mut (self.0 as *mut _));
-            }
-        }
-    }
-}
-
-/// RAII guard for `AVCodecContext` to ensure proper cleanup.
-struct AvCodecContextGuard(*mut AVCodecContext);
-
-impl AvCodecContextGuard {
-    unsafe fn new(codec: *const ff_sys::AVCodec) -> Result<Self, DecodeError> {
-        // SAFETY: Caller ensures codec pointer is valid
-        let codec_ctx = unsafe {
-            ff_sys::avcodec::alloc_context3(codec).map_err(|e| DecodeError::Ffmpeg {
-                code: e,
-                message: format!("Failed to allocate codec context: {e}"),
-            })?
-        };
-        Ok(Self(codec_ctx))
-    }
-
-    const fn as_ptr(&self) -> *mut AVCodecContext {
-        self.0
-    }
-
-    fn into_raw(self) -> *mut AVCodecContext {
-        let ptr = self.0;
-        std::mem::forget(self);
-        ptr
-    }
-}
-
-impl Drop for AvCodecContextGuard {
-    fn drop(&mut self) {
-        if !self.0.is_null() {
-            // SAFETY: self.0 is valid and owned by this guard
-            unsafe {
-                ff_sys::avcodec::free_context(&mut (self.0 as *mut _));
-            }
-        }
-    }
-}
+use crate::shared::guards_inner::{AvCodecContextGuard, AvFormatContextGuard};
 
 // ── ImageDecoderInner ─────────────────────────────────────────────────────────
